update migration, change OneToOne to ManyToOne in product-color.entity.ts and product-color-size.entity.ts, and add move images for product-color

+++ src/modules/admin/aws/aws.service.ts
@@ -2,6 +2,8 @@ import {Injectable} from "@nestjs/common"
 import * as AWS from "aws-sdk"
 import {ConfigService} from "@nestjs/config"
 import * as sharp from "sharp"
+import {ManagedUpload} from "aws-sdk/lib/s3/managed_upload"
+import SendData = ManagedUpload.SendData
 
 @Injectable()
 export class AwsService {
@@ -28,7 +30,7 @@ export class AwsService {
             quality?: number
             width?: number
         }
-    ) {
+    ): Promise<SendData & {name?: string}> {
         const imageBuffer = await sharp(file.buffer).webp({quality}).resize(width).toBuffer()
         const ext = "webp"
         const filename = Date.now() + "-" + Math.round(Math.random() * 1e9)
@@ -40,7 +42,8 @@ export class AwsService {
             ACL: "public-read"
         }
 
-        return this.aws.upload(params).promise()
+        const uploaded = await this.aws.upload(params).promise()
+        return {...uploaded, name: `${filename}.${ext}`}
     }
 
     async getFile(key: string) {
@@ -51,20 +54,29 @@ export class AwsService {
             }).promise()
     }
 
+    // TODO check and update
     async moveFile(keyFrom: string, keyTo: string) {
         // Copy file
-        await this.aws
-            .copyObject({
-                Bucket: this.configService.get("AWS_BUCKET_NANE"),
-                CopySource: this.configService.get("AWS_BUCKET_NANE") + "/" + keyFrom,
-                Key: keyTo
-            })
-            .promise()
-        // Delete file
-        this.aws.deleteObject({
-            Bucket: this.configService.get("AWS_BUCKET_NANE"),
-            Key: keyFrom
-        })
+        try {
+            const copyResult = await this.aws
+                .copyObject({
+                    Bucket: this.configService.get("AWS_BUCKET_NANE"),
+                    CopySource: this.configService.get("AWS_BUCKET_NANE") + "/" + keyFrom,
+                    Key: keyTo
+                })
+                .promise()
+
+            if (copyResult.CopyObjectResult) {
+                // Delete file
+                await this.aws.deleteObject({
+                    Bucket: this.configService.get("AWS_BUCKET_NANE"),
+                    Key: keyFrom
+                }).promise()
+            }
+        } catch (e) {
+            if (e.code === "NoSuchKey") console.log({code: e.code, keyFrom, keyTo})
+            console.log(e)
+        }
     }
 
     async deleteFile(key: string) {
